Log Dispatch

R6 Class that dispatches log messages throughout the application.

Details

This object is designed to a centralized logging dispatcher that renders log messages with the appropriate context of the calling object. The log_layout object is used to generate log message layouts (render formats), which are used by the LogDispatcher to render highly-customizable and detailed log messages.

See also

Methods

Public methods

LogDispatch$new()

LogDispatch$add_log_level()

LogDispatch$get_settings()

LogDispatch$set_settings()

Method new()

Creates a new instance of a log config.

Usage

LogDispatch$new()

Returns

A new LogLayout object.

Method add_log_level()

Adds dynamic function as a short-cut to log a message with a configured level.

Usage

LogDispatch$add_log_level(level)

Arguments

level

log level

Returns

reference to self to support chaining.

Method get_settings()

Public wrapper around logger settings.

Usage

LogDispatch$get_settings()

Returns

logger settings.

Method set_settings()

Public wrapper around updating settings.

Usage

LogDispatch$set_settings(settings)

Arguments

settings

Logger settings.

Returns

None.